Attend button now works correctly.
It seems to send 3 http requests to the server when attending,  not sure
why.  The attend sql query is now fixed so you dont get attended 3x
times.  Also removed the camera so dont start in the ocean any more. We
should maybe change this to our current location or Reykjavík :)